requirements.txt
asgiref==3.4.1
Django==4.0.1
django-phonenumber-field==6.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
phonenumberslite==8.12.41
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
